# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeamKnowledge API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  area?: string | null;
  createdAt: string;
}

export interface UserUpdate {
  name?: string;
  bio?: string;
  area?: string;
  avatarUrl?: string;
}

export interface Skill {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  tags: string[];
  usersCount?: number;
  createdAt: string;
}

export type UserSkillLevel = typeof UserSkillLevel[keyof typeof UserSkillLevel];


export const UserSkillLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
  expert: 'expert',
} as const;

export interface UserSkill {
  id: number;
  userId: number;
  skillId: number;
  level: UserSkillLevel;
  skill: Skill;
  user?: User;
  createdAt: string;
}

export interface SkillDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  tags: string[];
  usersCount?: number;
  createdAt: string;
  users: UserSkill[];
}

export interface SkillInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  tags?: string[];
}

export interface SkillUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  tags?: string[];
}

export type UserSkillInputLevel = typeof UserSkillInputLevel[keyof typeof UserSkillInputLevel];


export const UserSkillInputLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
  expert: 'expert',
} as const;

export interface UserSkillInput {
  skillId: number;
  level: UserSkillInputLevel;
}

export type UserSkillUpdateLevel = typeof UserSkillUpdateLevel[keyof typeof UserSkillUpdateLevel];


export const UserSkillUpdateLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
  expert: 'expert',
} as const;

export interface UserSkillUpdate {
  level: UserSkillUpdateLevel;
}

export type KnowledgeEntryType = typeof KnowledgeEntryType[keyof typeof KnowledgeEntryType];


export const KnowledgeEntryType = {
  article: 'article',
  link: 'link',
} as const;

export interface KnowledgeEntry {
  id: number;
  authorId: number;
  type: KnowledgeEntryType;
  title: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  description?: string | null;
  tags: string[];
  commentsCount: number;
  author?: User;
  createdAt: string;
  updatedAt?: string;
}

export type KnowledgeInputType = typeof KnowledgeInputType[keyof typeof KnowledgeInputType];


export const KnowledgeInputType = {
  article: 'article',
  link: 'link',
} as const;

export interface KnowledgeInput {
  type: KnowledgeInputType;
  /** @minLength 1 */
  title: string;
  content?: string;
  url?: string;
  description?: string;
  tags?: string[];
}

export interface KnowledgeUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  url?: string;
  description?: string;
  tags?: string[];
}

export type BoardEntryStatus = typeof BoardEntryStatus[keyof typeof BoardEntryStatus];


export const BoardEntryStatus = {
  learning: 'learning',
  doing: 'doing',
  done: 'done',
} as const;

export interface BoardEntry {
  id: number;
  authorId: number;
  title: string;
  /** @nullable */
  content?: string | null;
  status: BoardEntryStatus;
  tags?: string[];
  commentsCount: number;
  author?: User;
  createdAt: string;
  updatedAt?: string;
}

export type BoardInputStatus = typeof BoardInputStatus[keyof typeof BoardInputStatus];


export const BoardInputStatus = {
  learning: 'learning',
  doing: 'doing',
  done: 'done',
} as const;

export interface BoardInput {
  /** @minLength 1 */
  title: string;
  content?: string;
  status: BoardInputStatus;
  tags?: string[];
}

export type BoardUpdateStatus = typeof BoardUpdateStatus[keyof typeof BoardUpdateStatus];


export const BoardUpdateStatus = {
  learning: 'learning',
  doing: 'doing',
  done: 'done',
} as const;

export interface BoardUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  status?: BoardUpdateStatus;
  tags?: string[];
}

export interface Comment {
  id: number;
  authorId: number;
  content: string;
  author?: User;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  content: string;
}

export interface StatsOverview {
  totalUsers: number;
  totalSkills: number;
  totalKnowledge: number;
  totalBoards: number;
  totalComments: number;
}

export interface SkillStat {
  skillId: number;
  name: string;
  tags: string[];
  usersCount: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  board: 'board',
  knowledge: 'knowledge',
  comment: 'comment',
  skill: 'skill',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  authorId: number;
  authorName: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export type SkillMapEntrySkillsItem = {
  skillId: number;
  name: string;
  level: string;
};

export interface SkillMapEntry {
  userId: number;
  userName: string;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  area: string | null;
  skills: SkillMapEntrySkillsItem[];
}

export type ListSkillsParams = {
tag?: string;
search?: string;
};

export type ListKnowledgeParams = {
type?: ListKnowledgeType;
tag?: string;
search?: string;
authorId?: number;
};

export type ListKnowledgeType = typeof ListKnowledgeType[keyof typeof ListKnowledgeType];


export const ListKnowledgeType = {
  article: 'article',
  link: 'link',
} as const;

export type ListBoardsParams = {
userId?: number;
};

export type GetTopSkillsParams = {
limit?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};