# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TeamKnowledge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /knowledge:
    get:
      operationId: listKnowledge
      tags: [knowledge]
      summary: List all knowledge entries
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [article, link]
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: authorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Knowledge list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeEntry"
    post:
      operationId: createKnowledge
      tags: [knowledge]
      summary: Create a knowledge entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeEntry"
  /knowledge/{entryId}:
    get:
      operationId: getKnowledge
      tags: [knowledge]
      summary: Get a knowledge entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Knowledge entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeEntry"
        "404":
          description: Not found
    patch:
      operationId: updateKnowledge
      tags: [knowledge]
      summary: Update a knowledge entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeEntry"
    delete:
      operationId: deleteKnowledge
      tags: [knowledge]
      summary: Delete a knowledge entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Boards (aprendizados pessoais) ─────────────────────────────────────────
  /boards:
    get:
      operationId: listBoards
      tags: [boards]
      summary: List all boards (public feed)
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Boards list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BoardEntry"
    post:
      operationId: createBoard
      tags: [boards]
      summary: Create a board entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardEntry"
  /boards/{boardId}:
    get:
      operationId: getBoard
      tags: [boards]
      summary: Get a board entry
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Board entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardEntry"
        "404":
          description: Not found
    patch:
      operationId: updateBoard
      tags: [boards]
      summary: Update a board entry
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardEntry"
    delete:
      operationId: deleteBoard
      tags: [boards]
      summary: Delete a board entry
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Comments ───────────────────────────────────────────────────────────────
  /boards/{boardId}/comments:
    get:
      operationId: listBoardComments
      tags: [comments]
      summary: List comments on a board entry
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createBoardComment
      tags: [comments]
      summary: Add a comment to a board entry
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /knowledge/{entryId}/comments:
    get:
      operationId: listKnowledgeComments
      tags: [comments]
      summary: List comments on a knowledge entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createKnowledgeComment
      tags: [comments]
      summary: Add a comment to a knowledge entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Stats (wow endpoints) ──────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, name, email, createdAt]
      properties:
        id:            { type: integer }
        clerkId:       { type: string }
        name:          { type: string }
        email:         { type: string }
        avatarUrl:     { type: ["string", "null"] }
        bio:           { type: ["string", "null"] }
        area:          { type: ["string", "null"] }
        createdAt:     { type: string }
...
    KnowledgeEntry:
      type: object
      required: [id, authorId, type, title, tags, commentsCount, createdAt]
      properties:
        id:            { type: integer }
        authorId:      { type: integer }
        type:
          type: string
          enum: [article, link]
        title:         { type: string }
        content:       { type: ["string", "null"] }
        url:           { type: ["string", "null"] }
        description:   { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        commentsCount: { type: integer }
        author:
          $ref: "#/components/schemas/User"
        createdAt:     { type: string }
        updatedAt:     { type: string }
    KnowledgeInput:
      type: object
      required: [type, title]
      properties:
        type:
          type: string
          enum: [article, link]
        title:       { type: string, minLength: 1 }
        content:     { type: string }
        url:         { type: string }
        description: { type: string }
        tags:
          type: array
          items: { type: string }
    KnowledgeUpdate:
      type: object
      properties:
        title:       { type: string, minLength: 1 }
        content:     { type: string }
        url:         { type: string }
        description: { type: string }
        tags:
          type: array
          items: { type: string }

    # ── Board ─────────────────────────────────────────────────────────────────
    BoardEntry:
      type: object
      required: [id, authorId, title, status, commentsCount, createdAt]
      properties:
        id:            { type: integer }
        authorId:      { type: integer }
        title:         { type: string }
        content:       { type: ["string", "null"] }
        status:
          type: string
          enum: [learning, doing, done]
        tags:
          type: array
          items: { type: string }
        commentsCount: { type: integer }
        author:
          $ref: "#/components/schemas/User"
        createdAt:     { type: string }
        updatedAt:     { type: string }
    BoardInput:
      type: object
      required: [title, status]
      properties:
        title:   { type: string, minLength: 1 }
        content: { type: string }
        status:
          type: string
          enum: [learning, doing, done]
        tags:
          type: array
          items: { type: string }
    BoardUpdate:
      type: object
      properties:
        title:   { type: string, minLength: 1 }
        content: { type: string }
        status:
          type: string
          enum: [learning, doing, done]
        tags:
          type: array
          items: { type: string }

    # ── Comment ───────────────────────────────────────────────────────────────
    Comment:
      type: object
      required: [id, authorId, content, createdAt]
      properties:
        id:        { type: integer }
        authorId:  { type: integer }
        content:   { type: string }
        author:
          $ref: "#/components/schemas/User"
        createdAt: { type: string }
    CommentInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }

    # ── Stats ─────────────────────────────────────────────────────────────────
...