# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeamKnowledge API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Comment {
  id: number;
  authorId: number;
  content: string;
  author?: User;
  createdAt: string;
}