# lib/api-zod/src/generated/types/knowledgeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeamKnowledge API
 * OpenAPI spec version: 0.1.0
 */
import type { KnowledgeInputType } from './knowledgeInputType';

export interface KnowledgeInput {
  type: KnowledgeInputType;
  /** @minLength 1 */
  title: string;
  content?: string;
  url?: string;
  description?: string;
  tags?: string[];
}