# lib/api-zod/src/generated/types/knowledgeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeamKnowledge API
 * OpenAPI spec version: 0.1.0
 */

export interface KnowledgeUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  url?: string;
  description?: string;
  tags?: string[];
}