# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeamKnowledge API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  area?: string | null;
  createdAt: string;
}